Added OIDC-based user management to gptwol. This leads to the modifications to - wol.py - wol_form.html and subsequent modifications to - Dockerfile (app naming) - docker-publish.yml (image naming) - computers.txt (added users and flags) - requirements (added Authlib)
When running the container, make sure you set
these environment variables or have them in
oidc.txt:
	•	CLIENT_ID
	•	CLIENT_SECRET
	•	DISCOVERY_URL
	•	REDIRECT_URI
If all four are set, your app will use them
instead of reading oidc.txt. If any is missing,
the container will fall back to oidc.txt for the
missing values.

For every user in computers.txt, two optional
flags can be set.
- cannot_add_computer = true
- hide_details = true
Omitting a flag for a user has the same effect
of setting it as false. When hide_details is set
to true, only the computer name is shown to the
user so the user cannot see its IP and MAC.